+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+### Breaking Changes
+
+- `calcite-alert` - `open` and `close` methods have been removed. You can use the `active` prop to open or add an alert to the queue.
+- `calcite-alert` - `currentAlert` prop has been removed
+- `calcite-alert` - `alertQueue` prop has been removed (`queue` is emitted as a detail of the `calciteAlertOpen` and `calciteAlertClose` events)
+- `calcite-alert` - `alertQueueLength` prop has been removed
+
 ### Fixes
 
 - `calcite-modal` - turn off pointer events on hidden modals to prevent interaction (#549)
@@ -14,6 +21,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 - `calcite-modal` - add `background-color` property for light grey backgrounds (#527)
+- `calcite-alert` - `intlClose` prop has been added to optionally provide a translated override of the English "close" text
 
 ## [v1.0.0-beta.32]